# Revert Encoding/Decoding changes for better compatibility