--- a/parser_test.go
+++ b/parser_test.go
@@ -436,49 +436,49 @@ func TestParser_ParseUnverified(t *testing.T) {
 }
 
 var setPaddingTestData = []struct {
-	name    string
-	tokenString string
-	claims jwt.Claims
-	paddedDecode uint64
+	name          string
+	tokenString   string
+	claims        jwt.Claims
+	paddedDecode  bool
 	signingMethod jwt.SigningMethod
-	keyfunc jwt.Keyfunc
-	valid   bool
+	keyfunc       jwt.Keyfunc
+	valid         bool
 }{
 	{
-		name:    "Validated non-padded token with padding disabled",
-		tokenString: "",
-		claims:  jwt.MapClaims{"foo": "paddedbar"},
-		paddedDecode: jwt.DisablePadding,
+		name:          "Validated non-padded token with padding disabled",
+		tokenString:   "",
+		claims:        jwt.MapClaims{"foo": "paddedbar"},
+		paddedDecode:  false,
 		signingMethod: jwt.SigningMethodRS256,
-		keyfunc: defaultKeyFunc,
-		valid:   true,
+		keyfunc:       defaultKeyFunc,
+		valid:         true,
 	},
 	{
-		name:    "Validated non-padded token with padding enabled",
-		tokenString: "",
-		claims:  jwt.MapClaims{"foo": "paddedbar"},
-		paddedDecode: jwt.AllowPadding,
+		name:          "Validated non-padded token with padding enabled",
+		tokenString:   "",
+		claims:        jwt.MapClaims{"foo": "paddedbar"},
+		paddedDecode:  true,
 		signingMethod: jwt.SigningMethodRS256,
-		keyfunc: defaultKeyFunc,
-		valid:   true,
+		keyfunc:       defaultKeyFunc,
+		valid:         true,
 	},
 	{
-		name:    "Error for padded token with padding disabled",
-		tokenString: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJmb28iOiJwYWRkZWRiYXIifQ==.20kGGJaYekGTRFf8b0TwhuETcR8lv5z2363X5jf7G1yTWVTwOmte5Ii8L8_OQbYwPoiVHmZY6iJPbt_DhCN42AeFY74BcsUhR-BVrYUVhKK0RppuzEcSlILDNeQsJDLEL035CPm1VO6Jrgk7enQPIctVxUesRgswP71OpGvJxy3j1k_J8p0WzZvRZTe1D_2Misa0UDGwnEIHhmr97fIpMSZjFxlcygQw8QN34IHLHIXMaTY1eiCf4CCr6rOS9wUeu7P3CPkmFq9XhxBT_LLCmIMhHnxP5x27FUJE_JZlfek0MmARcrhpsZS2sFhHAiWrjxjOE27jkDtv1nEwn65wMw==",
-		claims:  jwt.MapClaims{"foo": "paddedbar"},
-		paddedDecode: jwt.DisablePadding,
+		name:          "Error for padded token with padding disabled",
+		tokenString:   "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJmb28iOiJwYWRkZWRiYXIifQ==.20kGGJaYekGTRFf8b0TwhuETcR8lv5z2363X5jf7G1yTWVTwOmte5Ii8L8_OQbYwPoiVHmZY6iJPbt_DhCN42AeFY74BcsUhR-BVrYUVhKK0RppuzEcSlILDNeQsJDLEL035CPm1VO6Jrgk7enQPIctVxUesRgswP71OpGvJxy3j1k_J8p0WzZvRZTe1D_2Misa0UDGwnEIHhmr97fIpMSZjFxlcygQw8QN34IHLHIXMaTY1eiCf4CCr6rOS9wUeu7P3CPkmFq9XhxBT_LLCmIMhHnxP5x27FUJE_JZlfek0MmARcrhpsZS2sFhHAiWrjxjOE27jkDtv1nEwn65wMw==",
+		claims:        jwt.MapClaims{"foo": "paddedbar"},
+		paddedDecode:  false,
 		signingMethod: jwt.SigningMethodRS256,
-		keyfunc: defaultKeyFunc,
-		valid:   false,
+		keyfunc:       defaultKeyFunc,
+		valid:         false,
 	},
 	{
-		name:    "Validated padded token with padding enabled",
-		tokenString: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJmb28iOiJwYWRkZWRiYXIifQ==.20kGGJaYekGTRFf8b0TwhuETcR8lv5z2363X5jf7G1yTWVTwOmte5Ii8L8_OQbYwPoiVHmZY6iJPbt_DhCN42AeFY74BcsUhR-BVrYUVhKK0RppuzEcSlILDNeQsJDLEL035CPm1VO6Jrgk7enQPIctVxUesRgswP71OpGvJxy3j1k_J8p0WzZvRZTe1D_2Misa0UDGwnEIHhmr97fIpMSZjFxlcygQw8QN34IHLHIXMaTY1eiCf4CCr6rOS9wUeu7P3CPkmFq9XhxBT_LLCmIMhHnxP5x27FUJE_JZlfek0MmARcrhpsZS2sFhHAiWrjxjOE27jkDtv1nEwn65wMw==",
-		claims:  jwt.MapClaims{"foo": "paddedbar"},
-		paddedDecode: jwt.AllowPadding,
+		name:          "Validated padded token with padding enabled",
+		tokenString:   "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJmb28iOiJwYWRkZWRiYXIifQ==.20kGGJaYekGTRFf8b0TwhuETcR8lv5z2363X5jf7G1yTWVTwOmte5Ii8L8_OQbYwPoiVHmZY6iJPbt_DhCN42AeFY74BcsUhR-BVrYUVhKK0RppuzEcSlILDNeQsJDLEL035CPm1VO6Jrgk7enQPIctVxUesRgswP71OpGvJxy3j1k_J8p0WzZvRZTe1D_2Misa0UDGwnEIHhmr97fIpMSZjFxlcygQw8QN34IHLHIXMaTY1eiCf4CCr6rOS9wUeu7P3CPkmFq9XhxBT_LLCmIMhHnxP5x27FUJE_JZlfek0MmARcrhpsZS2sFhHAiWrjxjOE27jkDtv1nEwn65wMw==",
+		claims:        jwt.MapClaims{"foo": "paddedbar"},
+		paddedDecode:  true,
 		signingMethod: jwt.SigningMethodRS256,
-		keyfunc: defaultKeyFunc,
-		valid:   true,
+		keyfunc:       defaultKeyFunc,
+		valid:         true,
 	},
 }
 
@@ -512,7 +512,7 @@ func TestSetPadding(t *testing.T) {
 			}
 
 		})
-		jwt.SetDecodePadding(jwt.DisablePadding)
+		jwt.SetDecodePadding(false)
 
 	}
 }