--- a/token.go
+++ b/token.go
@@ -17,15 +17,11 @@ const (
 var decodePaddingAllowed uint64
 
 // SetDecodePadding will switch the codec used for encoding/decoding JWTs respectively. Note that the JWS RFC7515
-// states that the tokens will utilize a Base64url encoding with no padding. Unfortuneately, some implementations
+// states that the tokens will utilize a Base64url encoding with no padding. Unfortunately, some implementations
 // of JWT are producing non-standard tokens, and thus require support for decoding. Note that this is a global
 // variable, and updating it will change the behavior on a package level.
 func SetDecodePadding(setPadding uint64) {
 	atomic.SwapUint64(&decodePaddingAllowed, setPadding)
-}
-
-func init() {
-	SetDecodePadding(DisablePadding)
 }
 
 // TimeFunc provides the current time when parsing token to validate "exp" claim (expiration time).