--- a/token.go
+++ b/token.go
@@ -8,7 +8,7 @@ import (
 )
 
 
-// DecodePaddingAllowed will switch the codec used for encoding/decoding JWTs respectively. Note that the JWS RFC7515
+// DecodePaddingAllowed will switch the codec used for decoding JWTs respectively. Note that the JWS RFC7515
 // states that the tokens will utilize a Base64url encoding with no padding. Unfortunately, some implementations
 // of JWT are producing non-standard tokens, and thus require support for decoding. Note that this is a global
 // variable, and updating it will change the behavior on a package level, and is also NOT go-routine safe.