… test/examplePaddedKey-public.pem