@ajermaky 2021-10-29
go test --bench=Bench --run=NONE .

Before Change:
`goos: linux
goarch: amd64
pkg: github.com/golang-jwt/jwt/v4
BenchmarkECDSAParsing/Basic_ES256-16         	  643087	      1725 ns/op	    1224 B/op	      39 allocs/op
BenchmarkECDSAParsing/Basic_ES384-16         	    1474	    761221 ns/op	 1737768 B/op	   14375 allocs/op
BenchmarkECDSAParsing/Basic_ES512-16         	    6580	    167854 ns/op	    1968 B/op	      44 allocs/op
BenchmarkECDSAParsing/basic_ES256_invalid:_foo_=>_bar-16         	  694071	      1691 ns/op	    1224 B/op	      39 allocs/op
BenchmarkECDSASigning/Basic_ES256-16                             	  330040	      3899 ns/op	    3871 B/op	      58 allocs/op
BenchmarkECDSASigning/Basic_ES256/sign-only-16                   	   44526	     27040 ns/op	    2946 B/op	      37 allocs/op
BenchmarkECDSASigning/Basic_ES384-16                             	    1507	    800376 ns/op	 1751594 B/op	   14474 allocs/op
BenchmarkECDSASigning/Basic_ES384/sign-only-16                   	     248	   4833601 ns/op	 1742866 B/op	   14392 allocs/op
BenchmarkECDSASigning/Basic_ES512-16                             	    5558	    182327 ns/op	    9157 B/op	      87 allocs/op
BenchmarkECDSASigning/Basic_ES512/sign-only-16                   	     772	   1551958 ns/op	    8114 B/op	      66 allocs/op
BenchmarkECDSASigning/basic_ES256_invalid:_foo_=>_bar-16         	  295987	      3931 ns/op	    3871 B/op	      58 allocs/op
BenchmarkHS256Signing-16                                                   	 1916628	       630.9 ns/op	    1554 B/op	      30 allocs/op
BenchmarkHS384Signing-16                                                   	 1563625	       757.1 ns/op	    1939 B/op	      30 allocs/op
BenchmarkHS512Signing-16                                                   	 1500824	       779.9 ns/op	    2035 B/op	      30 allocs/op
BenchmarkParseUnverified/map_claims-16                                     	 1519772	       797.8 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#01-16                                  	 1350411	       895.9 ns/op	    2328 B/op	      41 allocs/op
BenchmarkParseUnverified/map_claims#02-16                                  	 1339419	       903.0 ns/op	    2327 B/op	      41 allocs/op
BenchmarkParseUnverified/map_claims#03-16                                  	 1000000	      1004 ns/op	    2431 B/op	      46 allocs/op
BenchmarkParseUnverified/map_claims#04-16                                  	 1486888	       803.0 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#05-16                                  	 1501429	       802.0 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#06-16                                  	 1473259	       818.7 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#07-16                                  	 1459118	       822.5 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#08-16                                  	 1472215	       822.2 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#09-16                                  	 1484450	       815.4 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#10-16                                  	 1477605	       828.9 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#11-16                                  	 1479944	       812.6 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#12-16                                  	 1476196	       816.8 ns/op	    2192 B/op	      35 allocs/op
BenchmarkParseUnverified/standard_claims-16                                	 1412841	       841.5 ns/op	    2215 B/op	      35 allocs/op
BenchmarkParseUnverified/map_claims#13-16                                  	 1314297	       918.4 ns/op	    2328 B/op	      41 allocs/op
BenchmarkParseUnverified/map_claims#14-16                                  	 1307355	       932.0 ns/op	    2328 B/op	      41 allocs/op
BenchmarkParseUnverified/map_claims#15-16                                  	 1000000	      1034 ns/op	    2432 B/op	      46 allocs/op
BenchmarkParseUnverified/map_claims#16-16                                  	 1282791	       939.1 ns/op	    2328 B/op	      41 allocs/op
BenchmarkRSAParsing-16                                                     	  194726	      5517 ns/op	    9611 B/op	      93 allocs/op
BenchmarkRS256Signing-16                                                   	    8001	    150577 ns/op	   32934 B/op	     136 allocs/op
BenchmarkRS384Signing-16                                                   	    7653	    150839 ns/op	   33101 B/op	     136 allocs/op
BenchmarkRS512Signing-16                                                   	    7750	    154075 ns/op	   33193 B/op	     137 allocs/op
PASS
ok  	github.com/golang-jwt/jwt/v4	59.658s
`
With Change:
`goos: linux
goarch: amd64
pkg: github.com/golang-jwt/jwt/v4
BenchmarkECDSAParsing/Basic_ES256-16              624736              1731 ns/op            1224 B/op         39 allocs/op
BenchmarkECDSAParsing/Basic_ES384-16                1466            804266 ns/op         1737522 B/op      14373 allocs/op
BenchmarkECDSAParsing/Basic_ES512-16                5895            178290 ns/op            1968 B/op         44 allocs/op
BenchmarkECDSAParsing/basic_ES256_invalid:_foo_=>_bar-16                  726552              1751 ns/op            1224 B/op         39 allocs/op
BenchmarkECDSASigning/Basic_ES256-16                                      272566              4517 ns/op            3946 B/op         61 allocs/op
BenchmarkECDSASigning/Basic_ES256/sign-only-16                             45267             27159 ns/op            2970 B/op         38 allocs/op
BenchmarkECDSASigning/Basic_ES384-16                                        1388            814339 ns/op         1750131 B/op      14465 allocs/op
BenchmarkECDSASigning/Basic_ES384/sign-only-16                               244           4876130 ns/op         1746570 B/op      14422 allocs/op
BenchmarkECDSASigning/Basic_ES512-16                                        6546            188880 ns/op            9232 B/op         90 allocs/op
BenchmarkECDSASigning/Basic_ES512/sign-only-16                               739           1604532 ns/op            8138 B/op         67 allocs/op
BenchmarkECDSASigning/basic_ES256_invalid:_foo_=>_bar-16                  310046              4055 ns/op            3947 B/op         61 allocs/op
BenchmarkHS256Signing-16                                                         1865044               673.0 ns/op          1626 B/op         33 allocs/op
BenchmarkHS384Signing-16                                                         1482190               817.9 ns/op          2011 B/op         33 allocs/op
BenchmarkHS512Signing-16                                                         1302980               952.1 ns/op          2107 B/op         33 allocs/op
BenchmarkParseUnverified/map_claims-16                                           1407746               902.8 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#01-16                                        1223947               951.4 ns/op          2328 B/op         41 allocs/op
BenchmarkParseUnverified/map_claims#02-16                                        1288410               919.1 ns/op          2327 B/op         41 allocs/op
BenchmarkParseUnverified/map_claims#03-16                                        1000000              1067 ns/op            2495 B/op         48 allocs/op
BenchmarkParseUnverified/map_claims#04-16                                        1441771               837.4 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#05-16                                        1353867               897.7 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#06-16                                        1407392               828.7 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#07-16                                        1460398               835.1 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#08-16                                        1432087               853.3 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#09-16                                        1344019               864.0 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#10-16                                        1391701               881.8 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#11-16                                        1351614               877.9 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/map_claims#12-16                                        1341878               917.2 ns/op          2216 B/op         37 allocs/op
BenchmarkParseUnverified/standard_claims-16                                      1349334               856.1 ns/op          2215 B/op         35 allocs/op
BenchmarkParseUnverified/map_claims#13-16                                        1221648               955.7 ns/op          2328 B/op         41 allocs/op
BenchmarkParseUnverified/map_claims#14-16                                        1295176               930.1 ns/op          2328 B/op         41 allocs/op
BenchmarkParseUnverified/map_claims#15-16                                        1000000              1079 ns/op            2496 B/op         48 allocs/op
BenchmarkParseUnverified/map_claims#16-16                                        1314368               929.9 ns/op          2328 B/op         41 allocs/op
BenchmarkRSAParsing-16                                                            213618              5737 ns/op            9611 B/op         93 allocs/op
BenchmarkRS256Signing-16                                                            7338            152081 ns/op           32993 B/op        139 allocs/op
BenchmarkRS384Signing-16                                                            7332            157187 ns/op           33158 B/op        139 allocs/op
BenchmarkRS512Signing-16                                                            7813            156765 ns/op           33256 B/op        140 allocs/op
PASS
ok      github.com/golang-jwt/jwt/v4    61.993s
`